142. 정범호 / 몹 Dir문제 수정, Rot제대로 안되던 문제수정, Animation 특정프레임에 사운드 재생 쓰기쉽게 변경, Transform Move함수 네비메쉬 셀체크 return 제거, NPCBase 27개의 점 Pos추가

+++ TWEngine/Include/StrUtility.cpp
@@ -28,7 +28,7 @@ const string CStrUtility::TCHARToString(const TCHAR* _ptsz)
 const TCHAR* CStrUtility::StringToTCHAR(const string& _str)
 {
 	const char* all = _str.c_str();
-	int len = 1 + strlen(all);
+	int len = (int)(1 + strlen(all));
 	wchar_t* t = new wchar_t[len];
 	if (NULL == t) throw std::bad_alloc();
 	mbstowcs(t, all, len);
